MakaPakaApp/MakaPakaCourier/requirements.txt
flask
redis
pyopenssl
pyjwt
flask-jwt-extended 
authlib
requests